Refactored ntag to utilize Mui Tooltip. Refactored view model builder.
