+++ explorer/app/theme/definitions.d.ts
@@ -146,7 +146,6 @@ declare module "@mui/material/IconButton" {
 declare module "@mui/material/Paper" {
   interface PaperPropsVariantOverrides {
     footer: true;
-    ntag: true;
     panel: true;
   }
 }
